fix(bufferservice): Fixed exception when trying to heal while on cooldown
Config value was a Long but the annotated Seconds2MillisecondsConverter generates an Integer, which would cause a casting exception.
The config value is unboxed with longValue() in the datapack so it's safe to keep it as an Integer.

+++ src/main/java/com/l2jserver/gameserver/config/BufferServiceConfiguration.java
@@ -49,7 +49,7 @@ public interface BufferServiceConfiguration extends Reloadable {
 	
 	@Key("HealCooldown")
 	@ConverterClass(Seconds2MillisecondsConverter.class)
-	Long getHealCooldown();
+	Integer getHealCooldown();
 	
 	@Key("MaxUniqueLists")
 	Integer getMaxUniqueLists();
